Animação do Inimigo/Player feitas. Começando o patrulhamento do inimigo

+++ src/State.cpp
@@ -67,7 +67,7 @@ void State::FixedUpdateArray(float fixedDt)
 
 void State::UpdateArray(float dt)
 {
-	for (int i = 0; i < objectArray.size(); i++)
+	for (int i = 0; i < (int)objectArray.size(); i++)
 	{
 		objectArray[i]->Update(dt);
 	}
